src/main/webapp/app/entities/port/update/port-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { PortService } from '../service/port.service';
import { IPort, Port } from '../port.model';

import { PortUpdateComponent } from './port-update.component';

describe('Port Management Update Component', () => {
  let comp: PortUpdateComponent;
  let fixture: ComponentFixture<PortUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let portService: PortService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [PortUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(PortUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(PortUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    portService = TestBed.inject(PortService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const port: IPort = { id: 456 };

      activatedRoute.data = of({ port });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(port));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Port>>();
      const port = { id: 123 };
      jest.spyOn(portService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ port });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: port }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(portService.update).toHaveBeenCalledWith(port);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Port>>();
      const port = new Port();
      jest.spyOn(portService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ port });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: port }));
      saveSubject.complete();

      // THEN
      expect(portService.create).toHaveBeenCalledWith(port);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Port>>();
      const port = { id: 123 };
      jest.spyOn(portService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ port });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(portService.update).toHaveBeenCalledWith(port);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
